Further work on the grid
