+++ src/components/Prompt.js
@@ -1,14 +1,21 @@
 import React from 'react'
 import styled from 'styled-components'
-import Backdrop from './Backdrop'
+import Prompts from '../Prompts.json'
+import { getRandomInt } from '../utils/utils'
 
-export default function ({ children }) {
-  return (
-    <Backdrop>
-      <ButtonStyled>{children}</ButtonStyled>
-    </Backdrop>
-  )
+const prompt = JSON.stringify(
+  Prompts[getRandomInt(Prompts.length - 1)]['prompt']
+)
+
+export default function Prompt() {
+  return <PromptStyled>{prompt.slice(1, prompt.length - 1)}</PromptStyled>
 }
+
+const PromptStyled = styled.p`
+  margin-top: 20px;
+  margin-bottom: 20px;
+  -shadow: 20px 20px 60px #489dcf, -20px -20px 60px #62d5ff;
+`
 
 export const ButtonStyled = styled.button`
   height: 10%;
